Database: activated
Finally got around to adding the database, right now you can add stuff
to it and it'll save changes when you exit the program. I'll make it
more interactive when I make it more interactive. If you want to change
the stuff of items that're already inside the database, go to the txt
file and do it from there, but don't screw up the spacing.